requirements.txt
paho-mqtt
pymongo
fastapi
uvicorn
